requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pyyaml
openai==0.28
python-dotenv
Jinja2
    